Update data.toml with Modx Evolution
Added the PHP CMS Modx, which I use. This is the marketing description: “MODX Revolution is the web content management platform for those that truly care about no-compromise design and exceptional user experience. It gives you complete control over your site and content, with the flexibility and scalability to adapt to your changing needs.”