Add ponydocs to the list of awesome CMSes
Now pondering the proper plural of CMS. Ponydocs is a documentation-specific CMS that is, frankly, awesome for churning out documentation with different products, versions, child products, different manuals, and a generic "latest" for all your generic version needs. Supported and in heavy use by Splunk, it proves that a CMS can be a wiki but the content doesn't have to look like a wiki.